Make the initial polling follow the same code path as all other requests
Previously when we first started up, we tried to fetch an auth token
with an exponential backoff technique. Once that succeeded, we'd start
the normal polling for ads/points. However, that was a different
code path than the normal fetching of auth tokens based on getting 401s
from the server.

This patch changes it so that we try to get the ads/points with the
exponential backoff technique and once it succeeds once, we switch to
normal interval polling.